% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{politbaro}
\alias{politbaro}
\title{Politbarometer Extract}
\format{
A tibble:
}
\source{
\url{https://doi.org/10.4232/1.14479}
}
\usage{
politbaro
}
\description{
This dataset contains survey data from the Politbarometer study, cleaned for classroom use.
}
\keyword{datasets}
